testing:
- fixed unit testing / refactor
- install vitest coverage
