+++ package.json
@@ -45,6 +45,7 @@
   ],
   "devDependencies": {
     "@eslint/js": "^9.37.0",
+    "@vitest/coverage-v8": "^3.2.4",
     "eslint": "^9.37.0",
     "eslint-config-prettier": "^10.1.8",
     "eslint-plugin-prettier": "^5.5.4",
